chore(mapper): Merge LookupTable into MatcherTree
