+++ mapper/include/matchertree.h
@@ -1,28 +1,118 @@
 #pragma once
 #ifndef MATCHER_TREE_H_
 #define MATCHER_TREE_H_
-#include "lookup.h"
+#include "predicate.h"
 #include "llvm/ADT/SmallSet.h"
+#include "llvm/ADT/SmallVector.h"
 #include <map>
 #include <set>
 #include <string>
 
-struct LookupTable;
+struct Matcher {
+
+  // Matcher kind borrowed from DAGISelMatcher.h
+  // Keep updated with llvm's version
+  enum KindTy {
+    // Matcher state manipulation.
+    Scope,            // Push a checking scope.
+    RecordNode,       // Record the current node.
+    RecordChild,      // Record a child of the current node.
+    RecordMemRef,     // Record the memref in the current node.
+    CaptureGlueInput, // If the current node has an input glue, save it.
+    MoveChild,        // Move current node to specified child.
+    MoveParent,       // Move current node to parent.
+
+    // Predicate checking.
+    CheckSame,             // Fail if not same as prev match.
+    CheckChildSame,        // Fail if child not same as prev match.
+    CheckPatternPredicate, //
+    CheckPredicate,        // Fail if node predicate fails.
+    CheckOpcode,           // Fail if not opcode.
+    SwitchOpcode,          // Dispatch based on opcode.
+    CheckType,             // Fail if not correct type.
+    SwitchType,            // Dispatch based on type.
+    CheckChildType,        // Fail if child has wrong type.
+    CheckInteger,          // Fail if wrong val.
+    CheckChildInteger,     // Fail if child is wrong val.
+    CheckCondCode,         // Fail if not condcode.
+    CheckChild2CondCode,   // Fail if child is wrong condcode.
+    CheckValueType,
+    CheckComplexPat,
+    CheckAndImm,
+    CheckOrImm,
+    CheckImmAllOnesV,
+    CheckImmAllZerosV,
+    CheckFoldableChainNode,
+
+    // Node creation/emisssion.
+    EmitInteger,          // Create a TargetConstant
+    EmitStringInteger,    // Create a TargetConstant from a string.
+    EmitRegister,         // Create a register.
+    EmitConvertToTarget,  // Convert a imm/fpimm to target imm/fpimm
+    EmitMergeInputChains, // Merge together a chains for an input.
+    EmitCopyToReg,        // Emit a copytoreg into a physreg.
+    EmitNode,             // Create a DAG node
+    EmitNodeXForm,        // Run a SDNodeXForm
+    CompleteMatch,        // Finish a match and update the results.
+    MorphNodeTo,          // Build a node, finish a match and update results.
+
+    ScopeGroup,       // Custom: A child of a scope
+    SwitchTypeCase,   // Custom: A case of SwitchType
+    SwitchOpcodeCase, // Custom: A case of SwitchOpcode
+
+    // Highest enum value; watch out when adding more.
+    HighestKind = MorphNodeTo
+  };
+
+  size_t Begin = 0;
+  size_t End = 0;
+  KindTy Kind = Scope;
+  // To save space, this may either be a pattern or a pattern predicate index
+  size_t PIdx = 0;
+  std::string CaseName;
+
+  bool operator<(const Matcher &M) const;
+  bool operator==(const Matcher &N) const;
+  bool operator!=(const Matcher &N) const;
+
+  bool contains(size_t i) const;
+  bool contains(const Matcher &N) const;
+  bool hasPattern() const;
+  bool hasPatPred() const;
+  bool isLeaf() const;
+  bool isChild() const;
+  bool isCase() const;
+  size_t size() const;
+
+  static std::string getKindAsString(KindTy Kind);
+  std::string getKindAsString() const;
+};
 
 typedef std::vector<std::pair<size_t, size_t>> PatPredBlameList;
 typedef std::vector<std::pair<Matcher::KindTy, size_t>> MatcherBlameList;
 
+struct Pattern {
+  // We can but don't need to store them for now:
+  // std::string IncludePath;
+  // std::string PatternSrc;
+  llvm::SmallVector<size_t, 3> NamedPredicates;
+  size_t Index; // Index in the lookup table array
+};
+
 class MatcherTree {
-
-  const LookupTable &LT;
-  // The Matchers vector in lookup table for easy access
-  const std::vector<Matcher> &MT;
-
 public:
-  MatcherTree(const LookupTable &_LT) : LT(_LT), MT(_LT.Matchers) {}
-  MatcherTree() = delete;
+  std::vector<Pattern> Patterns;
+  size_t MatcherTableSize;
+  PredicateKeeper Predicates;
+  // Always sorted by matcher table index then size
+  std::vector<Matcher> Matchers;
+  size_t Verbosity = 0;
+
+  // Read from pattern lookup table JSON
+  MatcherTree(const std::string &Filename, bool NameCaseSensitive,
+              size_t Verbosity);
   MatcherTree(const MatcherTree &) = delete;
-  MatcherTree(MatcherTree &&M) : LT(M.LT), MT(M.MT) {}
+  MatcherTree(MatcherTree &&) = default;
 
   /// @brief Calculate matcher table coverage upper bound
   /// @return (covered indices, shadow map, pat pred idx -> coverage loss)
